requirements.txt
# YOLOv5 requirements
# Usage: pip install -r requirements.txt

# Base ----------------------------------------
# python>=3.8
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1  # Google Colab version
tqdm>=4.41.0
protobuf<=3.20.1  # https://github.com/ultralytics/yolov5/issues/8012
shapely>=2.0.4

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# pyqt5----------------------------------------
pyqt5>=5.15.9

# inpainting-------------------------------------
numba>=0.58.1

#package—---------------------------
pyinstaller>= 6.7.0

#torch torchvision torchaudio -cpu
torch==1.11.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.12.0+cpu --extra-index-url https://download.pytorch.org/whl/cpu
torchaudio==0.11.0 --extra-index-url https://download.pytorch.org/whl/cpu

#torch torchvision torchaudio -gpu cuda11.4
torch==1.11.0+cu114 --extra-index-url https://download.pytorch.org/whl/cu114
torchvision==0.12.0+cu114 --extra-index-url https://download.pytorch.org/whl/cu114
torchaudio==0.11.0 --extra-index-url https://download.pytorch.org/whl/cu114

